+++ CHANGELOG.md
@@ -140,11 +140,11 @@ Use a comma instead:
 
 - New configuration `up.fragment.config.renderOptions`. This is an object of default render options to always apply, even when not [navigating](/navigation).
 - `up.fragment.toTarget()` can now skip [target verifcation](/target-derivation#verification) by passing a `{ verify: false}` option.
-- New `[up-use-data]` attribute allows to [override data](/data#overriding) for the targeted fragment.
 - When calling `up.fragment.get()` with multiple search layers (e.g. `{ layer: "current, parent"}`), Unpoly will now search those layers in the given order.
 - Calling `up.fragment.get()` with an `Element`, that element is returned without further lookups.
-- The render option `{ useHungry }` has been renamed to `{ hungry }`. The corresponding HTML attribut is still `[up-use-hungry]` as to not conflict with `[up-hungry]`.
-- The render option `{ useKeep }` has been renamed to `{ keep }`. The corresponding HTML attribut is still `[up-use-keep]` as to not conflict with `[up-keep]`.
+- New `[up-use-data]` attribute allows to [override data](/data#overriding) for the targeted fragment. The corresponding render options is `{ data }`.
+- The render option `{ useHungry }` has been renamed to `{ hungry }`, but `{ useHungry }` is still accepted as an alias. The corresponding HTML attribut remains `[up-use-hungry]` as to not conflict with `[up-hungry]`.
+- The render option `{ useKeep }` has been renamed to `{ keep }`, but `{ useKeep }` is still accepted as an alias. The corresponding HTML attribut remains `[up-use-keep]` as to not conflict with `[up-keep]`.
 
 
 ### Layers