+++ CHANGELOG.md
@@ -8,6 +8,12 @@ If you're upgrading from an older Unpoly version you should load [`unpoly-migrat
 You may browse a formatted and hyperlinked version of this file at <https://unpoly.com/changes>.
 
 
+3.11.0
+------
+
+This release fixes an error in the minified Javascript (`unpoly.min.js`) (issue #703).
+
+
 3.10.0
 ------